# Optimize binary copy for ROS2 rcl uint8 array