--- a/config/config.exs
+++ b/config/config.exs
@@ -7,6 +7,7 @@ config :rclex,
     "std_msgs/msg/UInt8MultiArray",
     "std_msgs/msg/UInt32MultiArray",
     "geometry_msgs/msg/Twist",
+    "sensor_msgs/msg/Image",
     "sensor_msgs/msg/PointCloud",
     "diagnostic_msgs/msg/DiagnosticStatus",
     "action_msgs/msg/GoalInfo"