--- a/lib/mix/tasks/rclex/gen/msgs.ex
+++ b/lib/mix/tasks/rclex/gen/msgs.ex
@@ -157,12 +157,15 @@ defmodule Mix.Tasks.Rclex.Gen.Msgs do
     Enum.map_join(types, fn {:msg_type, type} ->
       function_prefix = Util.type_down_snake(type)
 
+      # WHY: Use ERL_NIF_DIRTY_JOB_IO_BOUND for destroy!, set!, and get!
+      #      These functions (de)allocate memory, and the size depends on the payload.
+      #      Therefore, the time required for the operations is unpredictable.
       """
       {"#{function_prefix}_type_support!", 0, nif_#{function_prefix}_type_support, REGULAR_NIF},
       {"#{function_prefix}_create!", 0, nif_#{function_prefix}_create, REGULAR_NIF},
-      {"#{function_prefix}_destroy!", 1, nif_#{function_prefix}_destroy, REGULAR_NIF},
-      {"#{function_prefix}_set!", 2, nif_#{function_prefix}_set, REGULAR_NIF},
-      {"#{function_prefix}_get!", 1, nif_#{function_prefix}_get, REGULAR_NIF},
+      {"#{function_prefix}_destroy!", 1, nif_#{function_prefix}_destroy, ERL_NIF_DIRTY_JOB_IO_BOUND},
+      {"#{function_prefix}_set!", 2, nif_#{function_prefix}_set, ERL_NIF_DIRTY_JOB_IO_BOUND},
+      {"#{function_prefix}_get!", 1, nif_#{function_prefix}_get, ERL_NIF_DIRTY_JOB_IO_BOUND},
       """
     end)
   end