- Replaced individual uint8 list copying with `memcpy` to avoid segmentation fault when handling large ERL_NIF_TERM arrays.
- This change also improves performance of the copy process.

This PR will close #347 .

---

TODO left

- [x] add comment, なぜ uint8 array を別扱いにするのか、 struct copy等
- [x] add bench
- [x] check test coverage
    - this should be fixed by #349 

---

## bench result

Before fix, it causes segmentation fault(#347), so I need to shrink the test image size like following. 
This bench uses this data as example image.

```elixir
%Rclex.Pkgs.SensorMsgs.Msg.Image{
  header: %Rclex.Pkgs.StdMsgs.Msg.Header{
    stamp: %Rclex.Pkgs.BuiltinInterfaces.Msg.Time{sec: 872_037, nanosec: 631_914},
    frame_id: "image"
  },
  height: 240,
  width: 320,
  encoding: "8UC1",
  is_bigendian: 0,
  step: 320 * 1,
  data: for(_ <- 1..(240 * 320 * 1), into: <<>>, do: <<:rand.uniform(256) - 1>>)
}
```

### before fix result (240 x 320 x 1)

```bash
mix run scripts/sensor_msgs_msg_image_benchee.exs
...
Name                 ips        average  deviation         median         99th %
create!/0         1.59 M        0.63 μs  ±7499.52%       0.179 μs        0.39 μs
destroy!/1        1.38 M        0.73 μs    ±98.33%        0.58 μs        2.20 μs
set!/2         0.00179 M      558.90 μs    ±37.47%      528.74 μs      898.12 μs
get!/1         0.00139 M      719.94 μs    ±16.27%      687.00 μs     1146.41 μs

Comparison: 
create!/0         1.59 M
destroy!/1        1.38 M - 1.15x slower +0.0958 μs
set!/2         0.00179 M - 886.86x slower +558.27 μs
get!/1         0.00139 M - 1142.38x slower +719.31 μs
```

### after fix result (240 x 320 x 1)

```bash
mix run scripts/sensor_msgs_msg_image_benchee.exs
...
Name                 ips        average  deviation         median         99th %
create!/0      1248.18 K        0.80 μs  ±5242.91%        0.31 μs        0.85 μs
destroy!/1      732.91 K        1.36 μs    ±39.05%        1.35 μs        2.16 μs
get!/1          236.84 K        4.22 μs   ±201.79%        3.96 μs        8.05 μs
set!/2          154.30 K        6.48 μs   ±701.11%        5.72 μs        8.71 μs

Comparison: 
create!/0      1248.18 K
destroy!/1      732.91 K - 1.70x slower +0.56 μs
get!/1          236.84 K - 5.27x slower +3.42 μs
set!/2          154.30 K - 8.09x slower +5.68 μs
```

### after fix result (480 x 640 x 3)

This bench uses a bigger image, (480 x 640 x 3).

```bash
mix run scripts/sensor_msgs_msg_image_benchee.exs
...
Name                 ips        average  deviation         median         99th %
create!/0      1329.89 K        0.75 μs  ±4950.56%        0.28 μs        0.92 μs
destroy!/1       42.81 K       23.36 μs    ±49.79%       21.69 μs       77.85 μs
get!/1           20.40 K       49.02 μs    ±32.07%       44.77 μs       98.30 μs
set!/2            4.64 K      215.59 μs    ±13.07%      208.50 μs      310.48 μs

Comparison: 
create!/0      1329.89 K
destroy!/1       42.81 K - 31.06x slower +22.61 μs
get!/1           20.40 K - 65.19x slower +48.27 μs
set!/2            4.64 K - 286.71x slower +214.84 μs
```