@pojiro 2024-10-09
@mlainez @FelixPe 

JFYI, I created this PR to fix #347. I think this PR modify the performance too.