@mlainez 2024-10-09
Hi @pojiro, I just tested this on nerves using an rpi4 with 2 USB cameras streaming in 720p at the same time and the performance improvement is **MASSIVE**!!! It seems close to what `usb_cam` achieves. With the `sensor_data` QoS profile, it streams both cams in 720p at a fairly decent framerate that I can work with. 

My 'non-scientific' tests below show an x3 improvement at 720p and significantly lower latency! At 640x480, since I limited the camera FPS to 30, I can't tell by how much, but performance is at least x3 as well.

Here are some results from `ros2 topic hz` after a few seconds using the `sensor_data` QoS profile and setting `CAP_PROP_FPS` in `Evision` to 30:

**No fix + 2 cams at 720p**
average rate: 3.552
	min: 0.189s max: 0.555s std dev: 0.08117s window: 333

**No fix + 2 cams at 640x480**
average rate: 9.512
	min: 0.065s max: 0.221s std dev: 0.03318s window: 338

**Binary fix + 2 cams at 720p**
average rate: 11.478
	min: 0.068s max: 0.124s std dev: 0.01384s window: 336

**Binary fix + 2 cams at 640x480**
average rate: 30.002
	min: 0.023s max: 0.045s std dev: 0.00314s window: 337

This PR rocks... :+1: