@takasehideki 2024-10-13
@pojiro thank you so much!! Of course creating a new release is also welcome if you want!